+++ src/main/java/com/example/fieldpasserbe/admin/service/AdminManageService.java
@@ -1,0 +1,7 @@
+package com.example.fieldpasserbe.admin.service;
+
+import com.example.fieldpasserbe.admin.vo.SimpleVO;
+
+public interface AdminManageService {
+    SimpleVO promoteAdmin(String email);
+}
