+++ src/main/java/com/example/fieldpasserbe/admin/service/impl/AdminServiceImpl.java
@@ -38,71 +38,6 @@ public class AdminServiceImpl implements AdminService {
     private final PunishService punishService;
     private final AdminRepositoryJPA adminRepository;
 
-    /**
-     * 관리자 로그인
-     * @param admin
-     * @param session
-     * @return
-     * @throws Exception
-     */
-    @Override
-    public AdminLoginVO adminLogin(AdminLoginRequestDTO admin, HttpSession session) throws Exception{
-
-        if (!isValidEmail(admin.getEmail())) {
-            throw new Exception("failed : 입력이 이메일 형식이 아닙니다.");
-        }
-
-        try {
-            Member member = memberService.findAdminByEmail(admin.getEmail()).get();
-            if (isValidPassword(admin, member)) {
-                session.setAttribute("email", member.getEmail());
-                memberService.updateVisitCount(member.getMemberId());
-                return AdminLoginVO.builder()
-                        .resultCode("success")
-                        .resultData(AdminLoginResponceDTO.builder()
-                                .email(member.getEmail())
-                                .profileImg(member.getProfileImg())
-                                .memberName(member.getMemberName())
-                                .build())
-                        .build();
-            } else {
-                throw new Exception("failed : 비밀번호가 틀렸습니다.");
-            }
-        } catch (NullPointerException e) {
-            throw new Exception("failed : 존재하지 않는 회원입니다.");
-        }
-    }
-
-    /**
-     * 비밀번호 매칭 검사
-     * @param admin
-     * @param member
-     * @return
-     */
-    private boolean isValidPassword(AdminLoginRequestDTO admin, Member member) {
-        if (member.getPassword().equals(admin.getPassword())) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * 이메일 형식 유효성 검사
-     * @param email
-     * @return
-     */
-    private static boolean isValidEmail(String email) {
-        String regex = "^[_a-z0-9-]+(.[_a-z0-9-]+)*@(?:\\w+\\.)+\\w+$";
-        Pattern p = Pattern.compile(regex);
-        Matcher m = p.matcher(email);
-        if(m.matches()) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
 
     /**
      * 전체 회원 조회
@@ -139,30 +74,6 @@ public class AdminServiceImpl implements AdminService {
                     .build();
         } catch (NullPointerException e) {
             throw new Exception("failed : 조회할 수 있는 회원이 없습니다.");
-        }
-    }
-
-    /**
-     * 일반 회원을 관리자로 등업
-     * @param email
-     * @return
-     */
-    @Override
-    public SimpleVO promoteAdmin(String email) {
-        try {
-            Admin newAdmin = Admin.builder()
-                    .member(memberService.findMemberByEmail(email).get())
-                    .promoteDate(LocalDateTime.now())
-                    .build();
-            newAdmin.promote();
-            adminRepository.save(newAdmin);
-            return SimpleVO.builder()
-                    .resultCode("success")
-                    .build();
-        } catch (NullPointerException e) {
-            return SimpleVO.builder()
-                    .resultCode("failed : 승격에 실패했습니다.")
-                    .build();
         }
     }
